Year 2 (2022-2023)/COMP0016 - Systems Engineering/Group Portfolio/code/src/motion-input-server/requirements.txt
psutil==5.9.4 
websockets==10.4
pyzmq==25.0.0 
Nuitka==1.5.3